+++ src/plugins/particles/objects/ParticleDisplay.cpp
@@ -136,14 +136,14 @@ void ParticleDisplay::particleColors(std::vector<Color>& output, ParticlePropert
 		// Check if all type IDs are within a small, non-negative range.
 		// If yes, we can use an array lookup strategy. Otherwise we have to use a dictionary lookup strategy, which is slower.
 		if(std::all_of(colorMap.begin(), colorMap.end(),
-				[&colorArray](const std::map<int,Color>::value_type& i) { return i.first >= 0 && i.first < colorArray.size(); })) {
+				[&colorArray](const std::map<int,Color>::value_type& i) { return i.first >= 0 && i.first < (int)colorArray.size(); })) {
 			colorArray.fill(defaultColor);
 			for(const auto& entry : colorMap)
 				colorArray[entry.first] = entry.second;
 			// Fill color array.
 			const int* t = typeProperty->constDataInt();
 			for(auto c = output.begin(); c != output.end(); ++c, ++t) {
-				if(*t >= 0 && *t < colorArray.size())
+				if(*t >= 0 && *t < (int)colorArray.size())
 					*c = colorArray[*t];
 				else
 					*c = defaultColor;
@@ -282,7 +282,7 @@ ParticlePrimitive::RenderingQuality ParticleDisplay::effectiveRenderingQuality(S
 	ParticlePrimitive::RenderingQuality renderQuality = renderingQuality();
 	if(renderQuality == ParticlePrimitive::AutoQuality) {
 		if(!positionProperty) return ParticlePrimitive::HighQuality;
-		int particleCount = positionProperty->size();
+		size_t particleCount = positionProperty->size();
 		if(particleCount < 2000 || renderer->isInteractive() == false)
 			renderQuality = ParticlePrimitive::HighQuality;
 		else if(particleCount < 100000)
@@ -307,7 +307,7 @@ void ParticleDisplay::render(TimePoint time, DataObject* dataObject, const Pipel
 	ParticlePropertyObject* transparencyProperty = ParticlePropertyObject::findInState(flowState, ParticleProperty::TransparencyProperty);
 
 	// Get number of particles.
-	int particleCount = positionProperty ? positionProperty->size() : 0;
+	int particleCount = positionProperty ? (int)positionProperty->size() : 0;
 
 	// Do we have to re-create the geometry buffer from scratch?
 	bool recreateBuffer = !_particleBuffer || !_particleBuffer->isValid(renderer);
